Address comments

+++ dynamallow/local.py
@@ -1,35 +1,36 @@
+import atexit
+import logging
 import os
 import random
 import socket
+import subprocess
 import tarfile
 import tempfile
-
-import subprocess
 
 try:
     from urllib import urlretrieve
 except ImportError:
     from urllib.request import urlretrieve
 
+log = logging.getLogger(__name__)
+
 
 class DynamoLocal(object):
     """
-    Spins up a local dynamo instance. This should ONLY be used for testing!! You must call
-    ``dynamo_local.shutdown()`` when you are done with the instance.
+    Spins up a local dynamo instance. This should ONLY be used for testing!! This instance
+    will register the cleanup method ``shutdown`` with the ``atexit`` module.
     """
-    def __init__(self, dynamo_dir, port=None, log=None):
-        self.port = port or self._get_random_port()
+    def __init__(self, dynamo_dir, port=None):
+        self.port = port or get_random_port()
         if not os.path.isdir(dynamo_dir):
-            if log:
-                log.info("Creating dynamo_local_dir: {0}".format(dynamo_dir))
+            log.info("Creating dynamo_local_dir: {0}".format(dynamo_dir))
             assert not os.path.exists(dynamo_dir)
             os.makedirs(dynamo_dir, 0o755)
 
         if not os.path.exists(os.path.join(dynamo_dir, 'DynamoDBLocal.jar')):
             temp_fd, temp_file = tempfile.mkstemp()
             os.close(temp_fd)
-            if log:
-                log.info("Downloading dynamo local to: {0}".format(temp_file))
+            log.info("Downloading dynamo local to: {0}".format(temp_file))
             urlretrieve(
                 'http://dynamodb-local.s3-website-us-west-2.amazonaws.com/dynamodb_local_latest.tar.gz',
                 temp_file
@@ -42,11 +43,10 @@ class DynamoLocal(object):
 
             os.unlink(temp_file)
 
-        if log:
-            log.info("Running dynamo from {dir} on port {port}".format(
-                dir=dynamo_dir,
-                port=self.port
-            ))
+        log.info("Running dynamo from {dir} on port {port}".format(
+            dir=dynamo_dir,
+            port=self.port
+        ))
         self.dynamo_proc = subprocess.Popen(
             (
                 'java',
@@ -60,21 +60,23 @@ class DynamoLocal(object):
             stderr=subprocess.PIPE,
             cwd=dynamo_dir
         )
+        atexit.register(self.shutdown)
 
     def shutdown(self):
-        self.dynamo_proc.terminate()
-        self.dynamo_proc.wait()
+        if self.dynamo_proc:
+            self.dynamo_proc.terminate()
+            self.dynamo_proc.wait()
+        self.dynamo_proc = None
 
-    @staticmethod
-    def _get_random_port():
-        """Find a random port that appears to be available"""
-        random_port = random.randint(25000, 55000)
-        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        try:
-            result = sock.connect_ex(('127.0.0.1', random_port))
-        finally:
-            sock.close()
-        if result == 0:
-            return DynamoLocal._get_random_port()
-        return random_port
 
+def get_random_port():
+    """Find a random port that appears to be available"""
+    random_port = random.randint(25000, 55000)
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    try:
+        result = sock.connect_ex(('127.0.0.1', random_port))
+    finally:
+        sock.close()
+    if result == 0:
+        return get_random_port()
+    return random_port
